scripts/requirements.txt
# Requirements for GitLab Reporter scripts
python-gitlab>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0
